requirements.txt
# Core dependencies
requests>=2.25.0
tqdm>=4.62.0
python-dotenv>=0.19.0
click>=8.0.0

# Optional dependencies
importlib-metadata>=1.0.0; python_version < "3.8"

# Development dependencies - uncomment for development
pytest>=6.0.0
pytest-cov>=2.12.0
pytest-mock>=3.6.0
black>=21.5b2
flake8>=3.9.0